+++ bin/commands/InspectCommand.js
@@ -8,9 +8,18 @@ const InspectCommand = {
     set target(value) {
         this._target = value;
     },
+
+    _targetUser: {},
+    get targetUser() {
+        return this._targetUser;
+    },
+    set targetUser(value) {
+        this._targetUser = value;
+    },
     
     inspect: async function(interaction) {
-        if (interaction.targetMember.user.bot) {
+        this.targetUser = interaction.options.get('target');
+        if (this.targetUser.bot) {
             await interaction.reply({
                 content: `${interaction.user} has targeted a bot! You can't target bots. Do something else.`,
                 files: [{
@@ -23,10 +32,10 @@ const InspectCommand = {
     },
 
     performInspect: async function(interaction) {
-        this.target = Players.find(interaction.targetUser.username);
+        this.target = Players.find(this.targetUser.username);
               // Send the interaction with the results
         await interaction.reply({
-            content: `${interaction.user} has brazenly inspected ${interaction.targetUser}. ${interaction.targetUser.username} currently has ${this.target.life>0 ? this.target.life : 'no'} hit points left.\nTheir stats:\nAttack Power: ${this.target.attack}\nDefense: ${this.target.defense}\nHeal Power: ${this.target.heal}`,
+            content: `${interaction.user} has brazenly inspected ${this.targetUser.member}. ${this.targetUser.user.username} currently has ${this.target.life>0 ? this.target.life : 'no'} hit points left.\nTheir stats:\nAttack Power: ${this.target.attack}\nDefense: ${this.target.defense}\nHeal Power: ${this.target.heal}`,
             files: [{
                 attachment: './assets/inspect.png'
             }]